Only define relocations if they aren't already
Currently, libdwarf/pro_init.c defines a number of relocation types that
are already defined by the Elf headers on illumos. This causes the
compiler to see a duplication definition, which it doesn't like. As
these are generally from the platform ABIs, only define them if they
aren't already defined.